craftx-stack/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
requests>=2.31.0
web3>=6.11.0
eth-account>=0.9.0
py-solc-x>=2.0.0
python-dotenv>=1.0.0
